[#90]refactor: scheduler 컴포넌트 분리 및 코드리뷰 반영
